requirements.txt
# ── Web Framework ──────────────────────────────────────
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
flask-session==0.6.0

# ── PDF & DOCX Parsing ─────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.0
pdfminer.six==20221105

# ── NLP ────────────────────────────────────────────────
nltk==3.8.1
spacy==3.7.2                   # UPGRADE 2: Named Entity Recognition

# ── Machine Learning (Original) ────────────────────────
scikit-learn==1.3.2
numpy==1.26.2

# ── UPGRADE 1: Semantic Similarity (BERT-based) ────────
sentence-transformers==2.7.0   # replaces TF-IDF for meaning-aware scoring
torch==2.2.0                   # required by sentence-transformers

# ── UPGRADE 3: Database ────────────────────────────────
flask-sqlalchemy==3.1.1        # ORM for database models
flask-login==0.6.3             # user authentication
flask-bcrypt==1.0.1            # password hashing

# ── UPGRADE 5: Charts & Visualisation ──────────────────
plotly==5.18.0                 # interactive charts
pandas==2.1.4                  # data manipulation for chart data

# ── Utilities ──────────────────────────────────────────
python-dotenv==1.0.0
gunicorn==21.2.0
